Create instagram_automation.py
This Python script utilizes the ppadb library to automate interactions with an Android device using ADB (Android Debug Bridge). The code connects to the ADB server, checks for connected devices, and selects the first connected device. It then defines a function mute_account to perform a sequence of touchscreen taps, muting posts and stories on a social media account. The script iterates through a list of accounts, muting each account in batches of 9, and swipes down to reveal more accounts. The total number of accounts to be processed is set to 1043, and the script calculates the number of rounds needed based on a batch size of 9.